Statkraft/requirements.txt
matplotlib==3.5.2
numpy==1.21.5
pandas==1.4.4
plotly==5.9.0
pygam==0.8.0
PyYAML==6.0
scipy==1.9.1
seaborn==0.11.2
statsmodels==0.13.2
tqdm==4.64.1
xgboost==1.7.3
